refactor: move ArgumentsContent components to own sub-dir
