+++ src/web/src/views/workspace/commandArgumentsContent/ArgNavBar.tsx
