+++ src/web/src/views/workspace/commandArgumentsContent/WSECArgumentSimilarPicker.tsx
